Add removal clause
We don't have a clear consensus on this so I am adding the most conservative option because I think it may be important to have *something* available in the event that a Member ceases to be active in the group. Language approximately mirrors that of the EC, except for "unanimous" in place of "75%" because functionally 75% rounds to 4 on a panel of 5.